chore: separate Paragraph component from BaseParagraph and fix dependencies
