+++ src/common/Paragraph.ts
@@ -1,6 +1,10 @@
-export interface ParagraphProps {
-  type: "paragraph"
+export interface BaseParagraphProps {
   content: string
+  className?: string
+}
+
+export interface ParagraphProps extends BaseParagraphProps {
+  type: "paragraph"
 }
 
 export default ParagraphProps
