+++ src/templates/next/components/Callout/Callout.tsx
@@ -1,10 +1,16 @@
 import { CalloutProps } from "~/common"
-import Paragraph from "../Paragraph"
+import BaseParagraph from "../shared/Paragraph"
+import { Paragraph } from "../../typography/Paragraph"
 
 const Callout = ({ content, variant }: CalloutProps) => {
   return (
-    <div className="bg-utility-info-subtle p-6 text-content-default text-lg leading-7 border border-utility-info rounded">
-      <Paragraph content={content} />
+    <div
+      className={`bg-utility-info-subtle p-6 border border-utility-info rounded`}
+    >
+      <BaseParagraph
+        content={content}
+        className={`text-content ${Paragraph[2]}`}
+      />
     </div>
   )
 }
