+++ src/templates/next/components/ContentPageHeader/ContentPageHeader.tsx
@@ -1,7 +1,8 @@
 import { ContentPageHeaderProps } from "~/common"
 import Button from "../Button"
 import Breadcrumb from "../Breadcrumb"
-import Paragraph from "../Paragraph"
+import { BaseParagraph } from "../shared/Paragraph"
+import { Paragraph } from "../../typography/Paragraph"
 
 const ContentPageHeader = ({
   title,
@@ -18,9 +19,10 @@ const ContentPageHeader = ({
           <h1 className="text-[2.75rem] leading-tight lg:text-[3.75rem] font-semibold text-content-strong">
             {title}
           </h1>
-          <div className="lg:text-xl lg:leading-8 text-content">
-            <Paragraph content={summary} />
-          </div>
+          <BaseParagraph
+            content={summary}
+            className={`text-content ${Paragraph[1]}`}
+          />
         </div>
         {buttonLabel && buttonUrl && (
           <Button
